requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
docutils==0.21.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.8.2
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
neo4j==5.27.0
nh3==0.2.20
openai==1.59.5
packaging==24.2
pkginfo==1.12.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
pytz==2024.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
SecretStorage==3.3.3
sniffio==1.3.1
tqdm==4.67.1
twine==6.0.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
